refactor by cr

+++ server/cluster/manager.go
@@ -290,16 +290,36 @@ func (m *managerImpl) GetTablesByShardIDs(clusterName, _ string, shardIDs []stor
 	return shardTables, nil
 }
 
+// DropTable is only used for the HTTP interface.
+// It only deletes the table data in ETCD and does not initiate a table deletion request to CeresDB.
 func (m *managerImpl) DropTable(ctx context.Context, clusterName, schemaName, tableName string) error {
 	cluster, err := m.getCluster(clusterName)
 	if err != nil {
 		return errors.WithMessage(err, "get cluster")
 	}
 
-	// TODO:
+	table, ok, err := cluster.GetMetadata().GetTable(schemaName, tableName)
+	if !ok {
+		return metadata.ErrTableNotFound
+	}
+	if err != nil {
+		return errors.WithMessage(err, "get table")
+	}
+
+	getShardNodeResult, err := cluster.GetMetadata().GetShardNodeByTableIDs([]storage.TableID{table.ID})
+	if err != nil {
+		return errors.WithMessage(err, "get shard node by tableID")
+	}
+	if len(getShardNodeResult.ShardNodes[table.ID]) != 1 || len(getShardNodeResult.Version) != 1 {
+		return metadata.ErrShardNotFound
+	}
+
+	shardID := getShardNodeResult.ShardNodes[table.ID][0].ID
+	version := getShardNodeResult.Version[shardID]
+
 	_, err = cluster.metadata.DropTable(ctx, schemaName, tableName, metadata.ShardVersionUpdate{
-		ShardID:     0,
-		PrevVersion: 0,
+		ShardID:     shardID,
+		PrevVersion: version,
 	})
 	if err != nil {
 		return errors.WithMessage(err, "cluster drop table")
